Moved RegisterForm into forms folder
